+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support for setting file capabilities via the RPMTAGS_FILECAPS header.
+- `PackageMetadata::get_file_entries` method can get capability headers for each file.
 
 ## 0.12.0